backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-debug-toolbar>=4.3.0
django-environ>=0.11.2
django-filter>=24.1
django-storages>=1.14.2
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary>=2.9.9
whitenoise==6.6.0
celery>=5.3.6
redis>=5.0.1
drf-yasg>=1.21.7
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
Pillow==10.2.0
requests==2.31.0 